Rework logs
